Server/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic_settings
python-multipart
opencv-python
numpy
Pillow
ultralytics
torch
torchvision
open-clip-torch
qdrant-client
python-dotenv
sqlalchemy
psycopg2-binary
neo4j
anyio
spacy
httpx
rich
loguru